Support 64-bit device id
Some users have expressed interest in being able to use 64-bit device
ids with bmv2. The biggest issue with this change is that the format for
notification messages (ageing, learning, port) has changes to accomodate
the larger device ids. As a consequence external tools subscribing to
these messages directly need to be updated.

+++ include/bm/bm_apps/learn.h
@@ -42,8 +42,8 @@ namespace bm_apps {
 class LearnListener {
  public:
   using buffer_id_t = uint64_t;
-  using switch_id_t = int;
-  using cxt_id_t = int;
+  using switch_id_t = uint64_t;
+  using cxt_id_t = uint32_t;
   using list_id_t = int;
 
   struct MsgInfo {
