+++ targets/simple_switch/tests/utils.cpp
@@ -83,8 +83,8 @@ void
 NNEventListener::receive_loop() {
   struct msg_hdr_t {
     int type;
-    int switch_id;
-    int cxt_id;
+    uint64_t switch_id;
+    uint32_t cxt_id;
     uint64_t sig;
     uint64_t id;
     uint64_t copy_id;
